flask_assignment/requirements.txt
cassandra-driver==3.25.0
click==8.0.1
dataclasses==0.8
Flask==2.0.1
geomet==0.2.1.post1
importlib-metadata==4.5.0
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
mysql-connector-python==8.0.25
numpy==1.19.5
pandas==1.1.5
protobuf==3.17.3
pymongo==3.11.4
python-dateutil==2.8.1
pytz==2021.1
six==1.16.0
typing-extensions==3.10.0.0
Werkzeug==2.0.1
zipp==3.4.1
